% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{bayesian_test_evaluation}
\alias{bayesian_test_evaluation}
\title{Not Finished 2
A/B Bayesian test of two proportions}
\usage{
bayesian_test_evaluation(
  data_input,
  alpha = 1,
  beta = 1,
  title_label = NULL,
  file_name = NULL,
  col_groups = c("blue", "red")
)
}
\arguments{
\item{data_input}{a data frame containing the variables in the model.}

\item{alpha}{priori parameters of the Beta distribution.}

\item{beta}{priori parameters of the Beta distribution.}

\item{col_groups}{a color vector used to color the groups.}
}
\description{
Not Finished 2
A/B Bayesian test of two proportions
}
\examples{
data(data_coin)
data_coin \%>\%
   bayesian_test_evaluation(
    alpha = 1, beta = 1,
    col_groups = c("Control" = "blue", "Variant 1" = "red"))

data(data_three_coins)
data_three_coins \%>\%
   bayesian_test_evaluation(
    alpha = 1, bet = 1,
    col_groups = c("Control" = "blue", "Variant 1" = "red", "Variant 2" = "green"))

mu <- .5
var <- .05
beta_parameters <- beta_moments_to_parameters(mu, mu * (1 - mu) * 0.05)

data_three_coins \%>\%
bayesian_test_evaluation(
   alpha = beta_parameters[1],
   bet = beta_parameters[2],
   col_groups = c("Control" = "blue", "Variant 1" = "red", "Variant 2" = "green"))
}
