% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{beta_moments_to_parameters}
\alias{beta_moments_to_parameters}
\title{Calculate the parameters of the Beta distribution when the mean and variance are known}
\usage{
beta_moments_to_parameters(mu, var)
}
\arguments{
\item{mu, var}{mean and variance  of the Beta distribution.}
}
\value{
Function returns the non-negative parameters of the Betadistribution
}
\description{
Calculate the parameters of the Beta distribution when the mean and variance are known
}
\examples{
# The uniform distribution(0,1) with mean 1/2 and variance 1/12
# is a Beta distribution with parameters alpha=1 and beta=1
beta_moments_to_parameters(1/2,1/12)

}
