% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{plot_bayesian_trend}
\alias{plot_bayesian_trend}
\title{Plot the evolution of the bayesian confidence interval}
\usage{
plot_bayesian_trend(
  data_input,
  alpha = 1,
  beta = 1,
  col_groups = c("blue", "red")
)
}
\arguments{
\item{data_input}{a data frame containing the variables in the model.}

\item{alpha, beta}{priori parameters of the Beta distribution.}

\item{col_groups}{a color vector used to color the groups.}
}
\value{
Returns a ggplot showing the evolution of the confidence intervaldd
}
\description{
Plot the evolution of the bayesian confidence interval
}
\examples{
library(ggplot2)
library(dplyr)
data(data_coins)
data_coin\%>\%
   plot_bayesian_trend(col_groups =c("Control"="blue","Variant 1"="red"))

data(data_three_coins)
data_three_coins \%>\%
   plot_bayesian_trend(col_groups = c("Control" = "blue", "Variant 1" = "red", "Variant 2" = "green"))
}
