% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{posterior_beta_dist}
\alias{posterior_beta_dist}
\title{Calculate the posterior parameters and confidence interval of the Beta distribution.}
\usage{
posterior_beta_dist(alpha = 1, beta = 1, x, n, conf_interval = 0.95)
}
\arguments{
\item{alpha, beta}{priori parameters of the Beta distribution.}

\item{x, n}{number of success and number of observations.}

\item{conf_interval}{Confidence interval}
}
\value{
returns the priori and posterior parameters, confidence interval of the Beta distribution.
}
\description{
Calculate the posterior parameters and confidence interval of the Beta distribution.
}
\examples{
# Posterior distrion when the prior is a Beta distribution with parameters alpha=1, beta=1,  and 30 success and 100 observations
posterior_beta_dist(1,1,30,100)
}
