% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{proportion_bayesian_test}
\alias{proportion_bayesian_test}
\title{Not Finished
A/B Bayesian test of two proportions}
\usage{
proportion_bayesian_test(
  data_input,
  alpha = 1,
  beta = 1,
  n_simulations = 100,
  col_groups = c("blue", "red")
)
}
\arguments{
\item{data_input}{a data frame containing the variables in the model.}

\item{alpha}{priori parameters of the Beta distribution.}

\item{beta}{priori parameters of the Beta distribution.}

\item{col_groups}{a color vector used to color the groups.}
}
\description{
Not Finished
A/B Bayesian test of two proportions
}
\examples{
data(data_coin)

data(data_coin)
data_coin \%>\%
   proportion_bayesian_test()

data(data_three_coins)

data_three_coins \%>\%
   proportion_bayesian_test(col_groups = c("Control" = "blue", "Variant 1" = "red", "Variant 2" = "green"))
}
